Fix the QD library support for hermite

+++ src/Hermite/neighbor.h
@@ -149,7 +149,7 @@ namespace H4 {
 
         //! check and add neighbor of single
         template <class Tp>
-        void checkAndAddNeighborSingle(const Float _r2, Tp& _particle, const Neighbor<Tparticle>& _nbp, const Float _index) {
+        void checkAndAddNeighborSingle(const Float _r2, Tp& _particle, const Neighbor<Tparticle>& _nbp, const int _index) {
             if (_r2<std::max(r_neighbor_crit_sq,_nbp.r_neighbor_crit_sq)) {
                 neighbor_address.addMember(NBAdr<Tparticle>(&_particle, _index));
                 n_neighbor_single++;
@@ -172,7 +172,7 @@ namespace H4 {
 
         //! check and add neighbor of group
         template <class Tgroup>
-        void checkAndAddNeighborGroup(const Float _r2, Tgroup& _group, const Float _index) {
+        void checkAndAddNeighborGroup(const Float _r2, Tgroup& _group, const int _index) {
             ASSERT(_r2>0.0);
             if (_r2<std::max(r_neighbor_crit_sq, _group.perturber.r_neighbor_crit_sq)) {
                 neighbor_address.addMember(NBAdr<Tparticle>(&_group.particles,_index));
